refactor(534): extract AudioPayload + trim files under 300-line limit
CI file length check on #559 flagged 4 files over 300 lines. Fixes:

- Extract AudioPayload class to new src/lyra/core/audio_payload.py.
  Runtime-import it in message.py (not TYPE_CHECKING) so get_type_hints()
  can resolve all InboundMessage annotations during NATS deserialization.
- Update 4 importers (tests/helpers/messages.py, tests/integration/
  test_voice_end_to_end.py, tests/nats/compat/test_inbound_audio_legacy.py,
  src/lyra/nats/compat/inbound_audio_legacy.py) to import from new path.
- Trim nats_bus.py: shorten NatsBus.inject() docstring + compact
  NatsBus.register() (removed redundant set-membership check since _registrations
  is already a set, and condensed Raises doc).
- Trim unified.py + bootstrap_wiring.py: condense compat-shim instantiation
  comments.

All four target files now ≤300 lines. 2466 tests pass, ruff clean.
