+++ src/lyra/core/message.py
@@ -5,6 +5,7 @@ from datetime import datetime, timezone
 from enum import Enum
 from typing import TYPE_CHECKING, Any, Literal
 
+from lyra.core.audio_payload import AudioPayload
 from lyra.core.trust import TrustLevel
 
 if TYPE_CHECKING:
@@ -137,22 +138,6 @@ class InboundAudio:
     is_mention: bool = False
     platform_meta: dict = field(default_factory=dict)
     routing: RoutingContext | None = None
-
-
-@dataclass(frozen=True)
-class AudioPayload:
-    """Audio payload nested inside InboundMessage when modality == 'voice'.
-
-    Replaces the parallel InboundAudio envelope in the unified inbound path
-    (issue #534). STT pipeline stage strips this field (sets to None) after
-    successful transcription to keep agent history free of raw audio bytes.
-    """
-
-    audio_bytes: bytes
-    mime_type: str
-    duration_ms: int | None = None
-    file_id: str | None = None
-    waveform_b64: str | None = None
 
 
 @dataclass
